Clean up the example
