+++ example/index.html
@@ -8,10 +8,15 @@
     />
 
     <title>React Baseline Inputs</title>
-    <link rel="stylesheet" href="./styles.scss" />
+    <link
+      rel="stylesheet"
+      href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
+      integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T"
+      crossorigin="anonymous"
+    />
   </head>
   <body>
     <div id="root"></div>
-    <script src="./src/App.tsx"></script>
+    <script src="./index.tsx"></script>
   </body>
 </html>
