+++ example/package.json
@@ -12,13 +12,11 @@
     "scheduler/tracing": "../node_modules/scheduler/tracing-profiling"
   },
   "dependencies": {
-    "bootstrap": "^4.5.0",
     "react-app-polyfill": "^1.0.1"
   },
   "devDependencies": {
     "@types/react": "^16.9.35",
     "@types/react-dom": "^16.9.8",
-    "parcel-bundler": "^1.12.4",
-    "sass": "^1.26.5"
+    "parcel-bundler": "^1.12.4"
   }
 }
