--- example/styles.scss
@@ -1,18 +1,0 @@
-@import "~bootstrap/scss/bootstrap";
-
-header {
-  margin-top: 2rem;
-  margin-bottom: 4rem;
-}
-
-section {
-  margin-bottom: 4rem;
-}
-
-.custom-switch {
-  .custom-control-input[aria-checked="true"] ~ .custom-control-label:after {
-    transform: translateX(
-      $custom-switch-width - $custom-control-indicator-size
-    );
-  }
-}
