+++ example/tsconfig.json
@@ -1,6 +1,5 @@
 {
   "extends": "../tsconfig.json",
-  "include": ["src"],
   "compilerOptions": {
     "rootDir": "../"
   }
